+++ telegrambot/nodes/out-node.js
@@ -19,6 +19,7 @@ module.exports = function (RED) {
     const path = require('path');
     const { pipeline } = require('stream');
     const fs = require('fs');
+    const QueueManager = require('../lib/queue-manager.js');
 
     // --------------------------------------------------------------------------------------------
     // The output node sends to the chat and passes the msg through.
@@ -44,8 +45,23 @@ module.exports = function (RED) {
         RED.nodes.createNode(this, config);
         let node = this;
         this.bot = config.bot;
+        this.messagesProcessed = 0;
+        this.retryDelayError429 = 3; // 3s when too many requests
+        this.retryDelayErrorNoConnection = 10; // 10s when not connected to internet
 
         let haserroroutput = config.haserroroutput || false;
+
+        // ------------------------------------------------------------------
+        // One queue per chatId to ensure that messages to the same chat are sent in order.
+        // Messages to different chats can be sent in parallel.
+        this.queueManager = new QueueManager();
+
+        this.enqueueMessage = function (chatId, msg, nodeSend, nodeDone) {
+            node.queueManager.enqueue(chatId, function () {
+                node.processMessage(chatId, msg, nodeSend, nodeDone);
+            });
+        };
+        // ------------------------------------------------------------------
 
         this.hasContent = function (msg) {
             let hasContent;
@@ -99,30 +115,72 @@ module.exports = function (RED) {
             return msg;
         };
 
-        this.processError = function (exception, msg, nodeSend, nodeDone) {
-            let errorMessage = 'Caught exception in sender node:\r\n' + exception + '\r\nwhen processing message: \r\n' + JSON.safeStringify(msg);
-
-            node.status({
-                fill: 'red',
-                shape: 'ring',
-                text: exception.message,
-            });
-
-            if (haserroroutput) {
-                let sendMessage = RED.util.cloneMessage(msg);
-                sendMessage.error = errorMessage;
-                nodeSend([null, sendMessage]);
+        this.processError = function (chatId, exception, msg, nodeSend, nodeDone) {
+            let retry = false;
+            let retryAfter = 10;
+            let retryReason = 'ERROR';
+            let error429 = String(exception).includes('Too Many Requests: retry after');
+            if (error429) {
+                retryReason = 'FLOODING';
+                retryAfter = exception.response.body.parameters.retry_after || node.retryDelayError429;
+                retry = true;
             } else {
-                if (nodeDone) {
-                    node.error(errorMessage, msg);
-                    nodeDone(errorMessage);
+                let errorNotFound = String(exception).includes('ENOTFOUND');
+                if (errorNotFound) {
+                    retryReason = 'ENOTFOUND';
+                    retryAfter = node.retryDelayErrorNoConnection;
+                    retry = true;
                 } else {
-                    node.error(errorMessage, msg);
+                    let errorConnectionReset = String(exception).includes('ECONNRESET');
+                    if (errorConnectionReset) {
+                        retryReason = 'ECONNRESET';
+                        retryAfter = node.retryDelayErrorNoConnection;
+                        retry = true;
+                    }
                 }
+            }
+
+            if (!retry) {
+                let errorMessage = 'Caught exception in sender node:\r\n' + exception + '\r\nwhen processing message: \r\n' + JSON.safeStringify(msg);
+
+                node.status({
+                    fill: 'red',
+                    shape: 'ring',
+                    text: exception.message,
+                });
+
+                if (haserroroutput) {
+                    let sendMessage = RED.util.cloneMessage(msg);
+                    sendMessage.error = errorMessage;
+                    nodeSend([null, sendMessage]);
+                } else {
+                    if (nodeDone) {
+                        node.error(errorMessage, msg);
+                        nodeDone(errorMessage);
+                    } else {
+                        node.error(errorMessage, msg);
+                    }
+                }
+            } else {
+                let errorMessage = retryReason + ': retrying in ' + retryAfter + 's';
+                node.status({
+                    fill: 'red',
+                    shape: 'ring',
+                    text: errorMessage,
+                });
+
+                node.queueManager.repeatProcessMessage(chatId, retryAfter);
             }
         };
 
-        this.processResult = function (result, msg, nodeSend, nodeDone) {
+        this.processResult = function (chatId, result, msg, nodeSend, nodeDone) {
+            node.messagesProcessed++;
+            node.status({
+                fill: 'green',
+                shape: 'ring',
+                text: 'messages sent: ' + node.messagesProcessed,
+            });
+
             if (result !== undefined) {
                 msg.payload.content = result;
                 msg.payload.sentMessageId = result.message_id;
@@ -132,6 +190,8 @@ module.exports = function (RED) {
             if (nodeDone) {
                 nodeDone();
             }
+
+            node.queueManager.processNext(chatId);
         };
 
         this.processMessage = function (chatId, msg, nodeSend, nodeDone) {
@@ -145,10 +205,10 @@ module.exports = function (RED) {
                 telegramBot
                     .forwardMessage(toChatId, chatId, messageId, msg.payload.forward.options)
                     .catch(function (ex) {
-                        node.processError(ex, msg, nodeSend, nodeDone);
+                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                     })
                     .then(function (result) {
-                        node.processResult(result, msg, nodeSend, nodeDone);
+                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                     });
             } else if (msg.payload.copy) {
                 // the message should be copied
@@ -158,10 +218,10 @@ module.exports = function (RED) {
                 telegramBot
                     .copyMessage(toChatId, chatId, messageId, msg.payload.copy.options)
                     .catch(function (ex) {
-                        node.processError(ex, msg, nodeSend, nodeDone);
+                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                     })
                     .then(function (result) {
-                        node.processResult(result, msg, nodeSend, nodeDone);
+                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                     });
             } else if (msg.payload.download) {
                 let fileId = msg.payload.download.fileId;
@@ -170,10 +230,10 @@ module.exports = function (RED) {
 
                 node.downloadFile(fileId, filePath, fileName)
                     .catch(function (ex) {
-                        node.processError(ex, msg, nodeSend, nodeDone);
+                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                     })
                     .then(function (result) {
-                        node.processResult(result, msg, nodeSend, nodeDone);
+                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                     });
             } else if (msg.payload.getfile) {
                 let fileId = msg.payload.getfile.fileId;
@@ -181,10 +241,10 @@ module.exports = function (RED) {
                 telegramBot
                     .getFile(fileId)
                     .catch(function (ex) {
-                        node.processError(ex, msg, nodeSend, nodeDone);
+                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                     })
                     .then(function (result) {
-                        node.processResult(result, msg, nodeSend, nodeDone);
+                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                     });
             } else {
                 if (msg.payload.type) {
@@ -213,7 +273,7 @@ module.exports = function (RED) {
                                     telegramBot
                                         .sendMessage(chatId, messageToSend, msg.payload.options || {})
                                         .then(function (result) {
-                                            node.processResult(result, msg, nodeSend, nodeDone);
+                                            node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                         })
                                         .catch(function (err) {
                                             // markdown error? try plain mode
@@ -232,14 +292,14 @@ module.exports = function (RED) {
                                                 telegramBot
                                                     .sendMessage(chatId, messageToSend, msg.payload.options || {})
                                                     .catch(function (ex) {
-                                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                                     })
                                                     .then(function (result) {
-                                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                                     });
                                                 return;
                                             } else {
-                                                node.processError(err, msg, nodeSend, nodeDone);
+                                                node.processError(chatId, err, msg, nodeSend, nodeDone);
                                             }
                                         });
                                 } while (!done);
@@ -251,10 +311,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendPhoto(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -275,10 +335,10 @@ module.exports = function (RED) {
                                     telegramBot
                                         .sendMediaGroup(chatId, msg.payload.content, msg.payload.options || {})
                                         .catch(function (ex) {
-                                            node.processError(ex, msg, nodeSend, nodeDone);
+                                            node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                         })
                                         .then(function (result) {
-                                            node.processResult(result, msg, nodeSend, nodeDone);
+                                            node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                         });
                                 } else {
                                     node.warn('msg.payload.content for mediaGroup is not an array of mediaItem');
@@ -290,7 +350,7 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendAudio(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
                                         msg.payload.content = result;
@@ -308,10 +368,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendDocument(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -321,10 +381,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendPoll(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.optional)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -334,10 +394,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendSticker(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -347,10 +407,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendDice(chatId, msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -360,10 +420,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendAnimation(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -373,10 +433,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendVideo(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -386,10 +446,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendVideoNote(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -399,10 +459,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendVoice(chatId, msg.payload.content, msg.payload.options || {}, msg.payload.fileOptions)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -412,10 +472,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendLocation(chatId, msg.payload.content.latitude, msg.payload.content.longitude, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -432,10 +492,10 @@ module.exports = function (RED) {
                                         msg.payload.options || {}
                                     )
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -451,10 +511,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendContact(chatId, msg.payload.content.phone_number, msg.payload.content.first_name, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -466,10 +526,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .editMessageLiveLocation(msg.payload.content.latitude, msg.payload.content.longitude, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -481,10 +541,10 @@ module.exports = function (RED) {
                             telegramBot
                                 .stopMessageLiveLocation(msg.payload.options)
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -506,10 +566,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .answerCallbackQuery(callbackQueryId, options)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -521,10 +581,10 @@ module.exports = function (RED) {
                             telegramBot
                                 .answerInlineQuery(msg.payload.inlineQueryId, msg.payload.results, msg.payload.options || {})
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -535,10 +595,10 @@ module.exports = function (RED) {
                             telegramBot
                                 .answerWebAppQuery(msg.payload.webAppQueryId, msg.payload.results, msg.payload.options || {})
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -549,10 +609,10 @@ module.exports = function (RED) {
                                 telegramBot
                                     .sendChatAction(chatId, msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -578,10 +638,10 @@ module.exports = function (RED) {
                         case 'deleteChatStickerSet':
                             telegramBot[type](chatId, msg.payload.options || {})
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             break;
 
@@ -593,10 +653,10 @@ module.exports = function (RED) {
                                 node.addChatIdToOptions(chatId, msg.payload.options);
                                 telegramBot[type](msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -607,10 +667,10 @@ module.exports = function (RED) {
                             if (this.hasContent(msg)) {
                                 node.editMessageMedia(msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -626,10 +686,10 @@ module.exports = function (RED) {
                             if (this.hasContent(msg)) {
                                 telegramBot[type](chatId, msg.payload.content)
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -651,10 +711,10 @@ module.exports = function (RED) {
                             if (this.hasContent(msg)) {
                                 telegramBot[type](chatId, msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -672,10 +732,10 @@ module.exports = function (RED) {
                             if (this.hasContent(msg)) {
                                 telegramBot[type](chatId, msg.payload.content, msg.payload.options || {})
                                     .catch(function (ex) {
-                                        node.processError(ex, msg, nodeSend, nodeDone);
+                                        node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                     })
                                     .then(function (result) {
-                                        node.processResult(result, msg, nodeSend, nodeDone);
+                                        node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                     });
                             }
                             break;
@@ -697,10 +757,10 @@ module.exports = function (RED) {
                                 msg.payload.options || {}
                             )
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -712,10 +772,10 @@ module.exports = function (RED) {
                             telegramBot
                                 .answerShippingQuery(msg.payload.shippingQueryId, msg.payload.ok, msg.payload.options || {})
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -727,10 +787,10 @@ module.exports = function (RED) {
                             telegramBot
                                 .answerPreCheckoutQuery(msg.payload.preCheckoutQueryId, msg.payload.ok, msg.payload.options || {})
                                 .catch(function (ex) {
-                                    node.processError(ex, msg, nodeSend, nodeDone);
+                                    node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                 })
                                 .then(function (result) {
-                                    node.processResult(result, msg, nodeSend, nodeDone);
+                                    node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                 });
                             //}
                             break;
@@ -749,10 +809,10 @@ module.exports = function (RED) {
                                 if (this.hasContent(msg)) {
                                     telegramBot[type](chatId, msg.payload.content, msg.payload.options || {})
                                         .catch(function (ex) {
-                                            node.processError(ex, msg, nodeSend, nodeDone);
+                                            node.processError(chatId, ex, msg, nodeSend, nodeDone);
                                         })
                                         .then(function (result) {
-                                            node.processResult(result, msg, nodeSend, nodeDone);
+                                            node.processResult(chatId, result, msg, nodeSend, nodeDone);
                                         });
                                 }
                             } else {
@@ -884,7 +944,7 @@ module.exports = function (RED) {
                 let telegramBot = this.config.getTelegramBot();
                 if (telegramBot) {
                     if (!Array.isArray(msg.payload.chatId)) {
-                        this.processMessage(msg.payload.chatId, msg, nodeSend, nodeDone);
+                        this.enqueueMessage(msg.payload.chatId, msg, nodeSend, nodeDone);
                     } else {
                         let chatIds = msg.payload.chatId;
                         let length = chatIds.length;
@@ -893,7 +953,7 @@ module.exports = function (RED) {
 
                             let clonedMsg = RED.util.cloneMessage(msg);
                             clonedMsg.payload.chatId = chatId;
-                            this.processMessage(chatId, clonedMsg, nodeSend, nodeDone);
+                            this.enqueueMessage(chatId, clonedMsg, nodeSend, nodeDone);
                         }
                     }
                 } else {